Fix manual signing/unsigning of integer values

+++ bans-core/src/main/java/space/arim/libertybans/core/database/JdbCaesarHelper.java
@@ -145,7 +145,7 @@ public class JdbCaesarHelper {
 		@Override
 		public Object adaptObject(Object parameter) {
 			if (parameter instanceof Long) {
-				return ((Long) parameter).longValue() + Long.MIN_VALUE;
+				return ((Long) parameter).longValue() + Integer.MIN_VALUE;
 			}
 			return parameter;
 		}
